fix: refacto Reactions
