+++ js/packages/components/chat/message/reactions/EmojiNumber.tsx
@@ -1,0 +1,13 @@
+import React from 'react'
+import { StyleProp, TextStyle, View } from 'react-native'
+
+import { UnifiedText } from '@berty/components/shared-components/UnifiedText'
+
+export const EmojiNumber: React.FC<{ count: number; style: StyleProp<TextStyle> }> = ({
+	count,
+	style,
+}) => (
+	<View>
+		<UnifiedText style={style}>{count}</UnifiedText>
+	</View>
+)
